Update clicker.py
Added a loop which waits for the user to check if the page has loaded or not. If the page is loaded, the user enters 'y' into the console and the program proceeds to the next function call. Each function call corresponds to the pages of the application. 

Still need to add:
-functions for the last 2 pages of the application
-If I can find a way to get the script to wait on its own until the page loads, it would mean the application is fully automated. (This is the ultimate goal for this script) Is the solution to go with the Selenium webdriver? There would be a lot more mouse clicks to automate in that situation.