Create AppPageClicker
This page is problemnatic for some reason. The script works for the majority of the page but as it gets further down, attempting to click the last box, the page hits an error and is forced to close. I'm trying several different solutions and combinations of clicks/key presses, nothing seems to work. 

Intel, you need to fix your website.